requirements.txt
langgraph>=0.2.0
google-genai
playwright>=1.44.0
python-dotenv>=1.0.0
cryptography>=42.0.0
aiofiles>=23.0.0
rich>=13.0.0
pydantic>=2.0.0
aiosqlite>=0.20.0
